BREAKING: Full EF rebuild
- Full EF rebuild. Migrator app in build/sql folder
- Slugs on URLs
- DateTimeOffset auditable fields
- Remove VC data repositories with DbContext
- Search and discovery improvements
- Randomized search with seed parameter
- Category facets
- Search to use discovery cards

+++ src/FoodStuffs.Model/Events/MealSets/GetMealSetResponsePantryIngredient.cs
@@ -2,4 +2,4 @@
 
 public record GetMealSetResponsePantryIngredient(
     string Name,
-    int Quantity);
+    decimal Quantity);
